overschool/requirements.txt
amqp==5.1.1
asgiref==3.8.1
opencv-python
async-timeout==4.0.2
atomicwrites==1.4.1
channels-redis==4.1.0
attrs==21.4.0
billiard==3.6.4.0
boto3==1.28.66
black==22.6.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
conditional==1.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
defusedxml==0.7.1
Deprecated==1.2.13
django-admin-rangefilter==0.12.0
Django==4.2.16
django-ckeditor==6.4.2
django-clone==3.0.6
django-cors-headers==3.13.0
django-embed-video==1.4.4
django-allauth==65.7.0
django-js-asset==2.0.0
django-pgconnection==1.0.2
django-pgtrigger==2.4.1
django-phonenumber-field==7.0.0
django-phonenumbers==1.0.1
django-redis==5.4.0
django-storages==1.12.3
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.4
environ==1.0
flake8==4.0.1
form==0.0.1
future==0.18.2
g4f
inflection==0.5.1
iniconfig==1.1.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
langdetect==1.0.9
MarkupSafe==2.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
oauthlib==3.2.0
packaging==21.3
pathspec==0.9.0
phonenumbers==8.12.56
Pillow==10.4.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
platformdirs==2.5.2
pluggy==1.0.0
pre-commit==3.3.1
prompt-toolkit==3.0.30
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.7.0
pyparsing==3.0.9
pytest==7.1.2
pytest-django==4.5.2
python-environ==0.4.54
python3-openid==3.2.0
pytz==2022.2.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
six==1.16.0
sentry-sdk==1.3.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.4.3
tomli==2.0.1
transliterate==1.10.2
types-cryptography==3.3.23
tzdata==2022.1
uritemplate==4.1.1
urllib3==2.0.7
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
daphne==4.0.0
channels==4.0.0
whitenoise==6.2.0
yadisk==1.3.3
pyTelegramBotAPI==4.17.0